[refactor] 커뮤니티 포스트 컴포넌트 분리
